mqtt-iwatch-wot/requirements/requirements.txt
numpy
scipy
paho-mqtt
